+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed `EditableTable` to `Table`, `EditableTable` kept for backwards compatibility and deprecated
 - Deprecation warning message for ModelList and EditableTable
 
+### Fixed
+- Cannot edit key on Dict fields in read_only mode.
+
 ## [0.4.0] - 2024-08-16
 ### Added
 - Possibility to set default repr type and kwargs directly in the pydantic model using repr_type and repr_kwargs